fastapi-asr-nemo/requirements.txt
fastapi
uvicorn
onnxruntime
torchaudio
torch
requests
nemo_toolkit[asr]